Component refactors and layout setup
