+++ src/components/Header/styles.module.css
@@ -1,0 +1,17 @@
+@import '../../styles/variables.css';
+
+.header {
+    background: var(--theme-color);
+    margin-bottom: 1.45rem;
+}
+
+.what {
+    margin: 0 auto;
+    max-width: 960;
+    padding: 1.45rem 1.0875rem;
+}
+
+.title {
+    color: var(--white);
+    text-decoration: none;
+}
