+++ src/components/Layout/index.js
@@ -1,0 +1,42 @@
+/**
+ * Layout component that queries for data
+ * with Gatsby's useStaticQuery component
+ *
+ * See: https://www.gatsbyjs.org/docs/use-static-query/
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useStaticQuery, graphql } from 'gatsby';
+
+import Header from '../Header';
+import '../../styles/reset.css';
+import styles from './styles.css';
+
+const Layout = ({ children }) => {
+  const data = useStaticQuery(graphql`
+    query SiteTitleQuery {
+      site {
+        siteMetadata {
+          title
+        }
+      }
+    }
+  `);
+
+  return (
+    <>
+      <Header siteTitle={data.site.siteMetadata.title} />
+      <div class="content">
+        <main className={styles.main}>{children}</main>
+      </div>
+      <footer className={styles.footer}>I am the footer</footer>
+    </>
+  );
+};
+
+Layout.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default Layout;
