+++ src/components/Layout/styles.css
@@ -1,0 +1,22 @@
+@import '../../styles/variables.css';
+
+html,
+body,
+#___gatsby,
+#gatsby-focus-wrapper {
+    height: 100%;
+}
+
+#gatsby-focus-wrapper {
+    display: flex;
+    flex-direction: column;
+}
+
+.content {
+    flex: 1 0 auto;
+    background-color: var(--white);
+}
+
+.footer {
+    flex-shrink: 0;
+}
